add clarification
Can we update the statement  in line 39 to below to calrify that the large audience size is evaluated only when group is added as recipient? Individual recipient counts greater that the large audience size does not trigger this mail tip.

"When sender adds a distribution group that has more members than the configured large audience size, they are shown the Large Audience MailTip."

This clarification is included in another article : https://docs.microsoft.com/en-us/exchange/clients-and-mobile-in-exchange-online/mailtips/mailtips#mailtips-in-exchange-online  but is missing in this article. This will help clear customer confusion in when they should expect this mail trip to trigger.